+++ CMakeLists.txt
@@ -183,6 +183,8 @@ set(FlatBuffers_Compiler_SRCS
   src/bfbs_gen_lua.h
   src/bfbs_gen_nim.h
   src/bfbs_namer.h
+  include/codegen/idl_namer.h
+  include/codegen/namer.h
   include/codegen/python.h
   include/codegen/python.cc
   include/flatbuffers/code_generators.h
